chore: rename hooks => hook
