+++ libs/_shared/hooks/convertToGithubImage.hook.ts
